+++ tests/TestCase/Command/GroupsCommandTest.php
@@ -1,0 +1,56 @@
+<?php
+/**
+ * This file is part of me-cms.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace Test\TestCase\Command;
+
+use Cake\Datasource\ModelAwareTrait;
+use MeCms\TestSuite\ConsoleIntegrationTestCase;
+
+/**
+ * GroupsCommandTest class
+ */
+class GroupsCommandTest extends ConsoleIntegrationTestCase
+{
+    use ModelAwareTrait;
+
+    /**
+     * Fixtures
+     * @var array
+     */
+    public $fixtures = [
+        'plugin.me_cms.Users',
+        'plugin.me_cms.UsersGroups',
+    ];
+
+    /**
+     * Test for `execute()` method
+     * @test
+     */
+    public function testExecute()
+    {
+        $this->loadModel('MeCms.UsersGroups');
+
+        $expectedRows = $this->UsersGroups->find()->map(function ($row) {
+            return [$row->id, $row->name, $row->label, $row->user_count];
+        });
+        $this->exec('me_cms.user groups');
+        $this->assertExitWithSuccess();
+        $this->assertTableHeadersEquals(['ID', 'Name', 'Label', 'Users']);
+        $this->assertTableRowsEquals($expectedRows->toList());
+
+        //Deletes all groups
+        $this->UsersGroups->deleteAll(['id >=' => '1']);
+        $this->exec('me_cms.user groups');
+        $this->assertExitWithError();
+        $this->assertErrorContains('There are no user groups');
+    }
+}